# docs/api.yaml
openapi: 3.0.3
info:
  title: WordWise Book Review Platform API
  description: |
    A comprehensive REST API for the WordWise book review platform. This API provides endpoints for user authentication, book management, reviews, ratings, recommendations, and user profiles.
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Book search and discovery
    - Review and rating system
    - AI-powered recommendations
    - User profile management
    - Popular books analytics
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited to 100 requests per 15 minutes per IP address in production.
    
    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "message": "Error description",
        "code": "ERROR_CODE",
        "details": "Additional error details"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: WordWise Development Team
    email: dev@wordwise.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.wordwise.com/api/v1
    description: Production server
...
paths:
...
  /reviews:
    post:
      tags:
        - Reviews
      summary: Create a new review
      description: Create a new book review
      operationId: createReview
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reviews/book/{bookId}:
    get:
      tags:
        - Reviews
      summary: Get reviews for a book
      description: Retrieve all reviews for a specific book
      operationId: getReviewsByBook
      parameters:
        - name: bookId
          in: path
          description: Book ID
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of reviews per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewsResponse'
  /reviews/user/{userId}:
    get:
      tags:
        - Reviews
      summary: Get reviews by user
      description: Retrieve all reviews by a specific user
      operationId: getReviewsByUser
      parameters:
        - name: userId
          in: path
          description: User ID
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of reviews per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewsResponse'
  /reviews/{id}:
    get:
      tags:
        - Reviews
      summary: Get review by ID
      description: Retrieve a specific review by its ID
      operationId: getReviewById
      parameters:
        - name: id
          in: path
          description: Review ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Reviews
      summary: Update review
      description: Update an existing review
      operationId: updateReview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Review ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not the review owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Reviews
      summary: Delete review
      description: Delete an existing review
      operationId: deleteReview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Review ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not the review owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: "Error description"
            code:
              type: string
              example: "ERROR_CODE"
            details:
              type: string
              example: "Additional error details"
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
...
    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          example: "The Great Gatsby"
        author:
          type: string
          example: "F. Scott Fitzgerald"
        description:
          type: string
          example: "A classic American novel..."
        coverImageUrl:
          type: string
          format: uri
          example: "https://example.com/cover.jpg"
        genres:
          type: array
          items:
            type: string
          example: ["Fiction", "Classic Literature"]
        publishedYear:
          type: integer
          example: 1925
        averageRating:
          type: number
          format: decimal
          example: 4.2
        reviewCount:
          type: integer
          example: 150
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        bookId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        content:
          type: string
          example: "This is a great book!"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
        user:
          $ref: '#/components/schemas/User'
        book:
          $ref: '#/components/schemas/Book'
    ReviewResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Review'
    ReviewsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            pagination:
              $ref: '#/components/schemas/Pagination'
    CreateReviewRequest:
      type: object
      required:
        - bookId
        - rating
      properties:
        bookId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        content:
          type: string
          maxLength: 2000
          example: "This is a great book!"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
    UpdateReviewRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 2000
          example: "Updated review content"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false